requirements.txt
flask==3.0.0
anthropic==0.40.0
Pillow==10.2.0
colorthief==0.2.1
python-dotenv==1.0.0
requests==2.31.0